refactor: Create widget folder for all widget files
