--- plugins/ui/src/js/src/ReactPanelManagerProvider.tsx
@@ -1,30 +1,0 @@
-import React, { useMemo } from 'react';
-import {
-  ReactPanelManager,
-  ReactPanelManagerContext,
-} from './layout/ReactPanelManager';
-
-export function ReactPanelManagerProvider({
-  children,
-  metadata,
-  onOpen,
-  onClose,
-  getPanelId,
-}: React.PropsWithChildren<ReactPanelManager>): JSX.Element {
-  const manager = useMemo(
-    () => ({
-      metadata,
-      onOpen,
-      onClose,
-      getPanelId,
-    }),
-    [metadata, onOpen, onClose, getPanelId]
-  );
-  return (
-    <ReactPanelManagerContext.Provider value={manager}>
-      {children}
-    </ReactPanelManagerContext.Provider>
-  );
-}
-
-export default ReactPanelManagerProvider;
