+++ plugins/ui/src/js/src/widget/WidgetHandler.test.tsx
