+++ plugins/ui/src/js/src/widget/WidgetHandler.tsx
@@ -22,7 +22,7 @@ import {
   isCallableNode,
   isElementNode,
   isObjectNode,
-} from './elements/ElementUtils';
+} from '../elements/ElementUtils';
 import { WidgetMessageEvent, WidgetWrapper } from './WidgetTypes';
 import DocumentHandler from './DocumentHandler';
 import { getComponentForElement } from './WidgetUtils';
